refactored utils projects so fakeIngestion can be independent
