+++ test/smoke/build.gradle
@@ -8,6 +8,15 @@ allprojects {
 	}
 }
 
+subprojects {
+	ext {
+		// FIXME FEATURE use jars that we provide
+		aiAgentJar = project(':agent')
+		aiCoreJar = project(':core')
+		aiWebJar =  project(':web')
+	}
+}
+
 clean {
 	delete file(sharedOutputRoot)
 }
