+++ test/smoke/framework/fakeIngestion/standalone/build.gradle
@@ -1,0 +1,17 @@
+apply plugin: 'java'
+
+repositories {
+	jcenter()
+	mavenCentral()
+}
+
+dependencies {
+	compile 'com.google.guava:guava:23.0'
+	compile 'junit:junit:4.12'
+	compile 'org.eclipse.jetty.aggregate:jetty-all:9.4.7.v20170914'
+	compile 'com.google.code.gson:gson:2.8.2'
+	compile aiCoreJar
+	compile project(':test:smoke:framework:fakeIngestion:servlet')
+}
+
+compileJava.sourceCompatibility = 1.8
