+++ test/smoke/framework/testCases/src/main/java/com/microsoft/applicationinsights/smoketest/FixedAiTestCases.java
@@ -3,7 +3,6 @@ package com.microsoft.applicationinsights.smoketest;
 import com.google.common.base.Preconditions;
 import com.microsoft.applicationinsights.TelemetryClient;
 import com.microsoft.applicationinsights.telemetry.Duration;
-import com.microsoft.applicationinsights.telemetry.RequestTelemetry;
 import com.microsoft.applicationinsights.telemetry.SeverityLevel;
 
 import java.util.Date;
