+++ test/smoke/framework/testCore/build.gradle
@@ -1,0 +1,15 @@
+apply plugin: 'java-library'
+
+repositories {
+	jcenter()
+}
+
+dependencies {
+	compile 'com.google.guava:guava:23.0'
+	compile 'junit:junit:4.12'
+	compile aiCoreJar
+	compile project(':test:smoke:framework:utils')
+	compile project(':test:smoke:framework:fakeIngestion:standalone')
+}
+
+compileJava.sourceCompatibility = 1.8
