+++ test/smoke/framework/testCore/src/main/java/com/microsoft/applicationinsights/smoketest/ExpectSomeTelemetry.java
