+++ test/smoke/framework/testCore/src/main/java/com/microsoft/applicationinsights/smoketest/docker/ContainerInfo.java
