+++ test/smoke/framework/utils/build.gradle
@@ -5,13 +5,9 @@ repositories {
 }
 
 dependencies {
-	compile 'com.google.guava:guava:23.0'
-	compile 'junit:junit:4.12'
-	compile 'org.eclipse.jetty.aggregate:jetty-all:9.4.7.v20170914'
-	compile 'javax.servlet:javax.servlet-api:3.0.1'
 	compile 'com.google.code.gson:gson:2.8.2'
     compile 'org.apache.httpcomponents:httpclient:4.5.3'
-	compile project(':core')
+	compile aiCoreJar
 }
 
 compileJava.sourceCompatibility = 1.8
