+++ test/smoke/framework/utils/src/test/java/com/microsoft/applicationinsights/smoketest/JsonHelperTests.java
