+++ test/smoke/testApps/PerfTestApp/build.gradle
@@ -5,8 +5,8 @@ dependencies {
     compile aiWebJar
     compile 'com.google.guava:guava:20.0'
 
-    compile project(':test:smoke:utils')
-    
+    compile project(':test:smoke:framework:testCases')
+
     providedCompile 'javax.servlet:javax.servlet-api:3.0.1'
 
     providedRuntime 'mysql:mysql-connector-java:5.1.44'
