+++ test/smoke/testApps/PerfTestApp/src/main/java/com/microsoft/applicationinsights/testapps/perf/servlets/TrackServlet.java
@@ -2,7 +2,6 @@ package com.microsoft.applicationinsights.testapps.perf.servlets;
 
 import com.microsoft.applicationinsights.TelemetryClient;
 import com.microsoft.applicationinsights.smoketest.FixedAiTestCases;
-import com.microsoft.applicationinsights.testapps.perf.TestCaseRunnable;
 import com.microsoft.applicationinsights.testapps.perf.TestCaseRunnableFactory;
 
 import javax.servlet.annotation.WebServlet;
